+++ Data Structures/SegTreeRec.cpp
@@ -8,32 +8,32 @@ class SegTree { //Tudo 0-based
 
 public :
  
-    void build(int node, int L, int R, vector<T> &V) {
+    void build(size_t node, size_t L, size_t R, vector<T> &V) {
         if (L == R) Tree[node] = V[L];
         else {
-            int M = (L + R) / 2;
+            size_t M = (L + R) / 2;
             build(2 * node, L, M, V);
             build(2 * node + 1, M + 1, R, V);
             Tree[node] = op(Tree[2 * node], Tree[2 * node + 1]);
         }
     }
 
-    T query(int l, int r, int node, int L, int R) {
+    T query(size_t l, size_t r, size_t node, size_t L, size_t R) {
         if (l > R or r < L)
             return neutral;
         if (l <= L and R <= r)
             return Tree[node];
         
-        int M = (L + R) / 2;
+        size_t M = (L + R) / 2;
         return op(query(l, r, 2 * node, L, M), query(l, r, 2 * node + 1, M + 1, R));
     }
 
-    T query(int l, int r) { return query(l, r, 1, 0, sz - 1); }
+    T query(size_t l, size_t r) { return query(l, r, 1, 0, sz - 1); }
 
-    void update(int id, T val, int node, int L, int R) {
+    void update(size_t id, T val, size_t node, size_t L, size_t R) {
         if (L == R) Tree[node] = op(Tree[node], val);
         else {
-            int M = (L + R) / 2;
+            size_t M = (L + R) / 2;
             
             if (id <= M)
                 update(id, val, 2 * node, L, M);
@@ -44,7 +44,7 @@ public :
         }
     }
 
-    void update(int id, T val) { update(id, val, 1, 0, sz - 1); }
+    void update(size_t id, T val) { update(id, val, 1, 0, sz - 1); }
 
 	SegTree(const vector<T> &V, function<T(T,T)> op_, T neutral_) : neutral(neutral_) , op(op_) {
 		sz = V.size();
@@ -53,13 +53,13 @@ public :
         build(1, 0, sz - 1, V);
     }
 
-    SegTree(int sz_, function<T(T, T)> op_, T neutral_) : sz(sz_) , neutral(neutral_) , op(op_) {
+    SegTree(size_t sz_, function<T(T, T)> op_, T neutral_) : sz(sz_) , neutral(neutral_) , op(op_) {
         Tree = vector<T>(3 * sz, neutral);
     }
  
 private :
 
-    int sz;
+    size_t sz;
     vector<T> Tree;
     const T neutral;
     function<T(T, T)> op;
